+++ gateway/src/config.rs
@@ -10,6 +10,8 @@ pub struct Config {
     #[serde(default)]
     pub auth: Auth,
     #[serde(default)]
+    pub manage_auth: Auth,
+    #[serde(default)]
     pub log: Log,
     #[serde(default)]
     pub reforward: Reforward,
