+++ gateway/src/route/manager.rs
@@ -1,0 +1,7 @@
+use axum::Router;
+
+use crate::AppState;
+
+pub fn route() -> Router<AppState> {
+    Router::new()
+}
