+++ gateway/src/route/mod.rs
@@ -1,0 +1,4 @@
+pub mod hook;
+pub mod manager;
+pub mod proxy;
+pub mod r#static;
